+++ src/main/java/com/ronreynolds/smartsheet/api/util/Sheets.java
@@ -1,19 +1,23 @@
 package com.ronreynolds.smartsheet.api.util;
 
+import com.ronreynolds.smartsheet.ApiClient;
 import com.ronreynolds.smartsheet.ApiException;
+import com.ronreynolds.smartsheet.api.RowsApi;
 import com.ronreynolds.smartsheet.api.SheetsApi;
 import com.ronreynolds.smartsheet.model.Column;
+import com.ronreynolds.smartsheet.model.Row;
 import com.ronreynolds.smartsheet.model.Sheet;
 import com.ronreynolds.smartsheet.model.SheetExclude;
 import com.ronreynolds.smartsheet.model.SheetInclude;
 import com.ronreynolds.smartsheet.model.SheetListingDataInner;
+import com.ronreynolds.util.assertions.State;
 import lombok.NonNull;
 
 import java.time.OffsetDateTime;
-import java.util.Date;
 import java.util.EnumSet;
 import java.util.List;
-import java.util.Objects;
+import java.util.function.Consumer;
+import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
 /**
@@ -22,55 +26,20 @@ import java.util.stream.Collectors;
  */
 @SuppressWarnings("unused")
 public class Sheets {
-    static String LINE_DELIMITER = "";    // no line-breaks by default
-
-/*
-    @NonNull
-    public static CharSequence toString(@NonNull Sheet sheet) {
-        try {
-            return toString(sheet, null);
-        } catch (ApiException e) {
-            throw new IllegalStateException("should be impossible", e);
-        }
-    }
-
-    @NonNull
-    public static CharSequence toString(@NonNull Sheet sheet, SheetsApi client, ToStringOptions... options)
-            throws ApiException {
-        StringBuilder buf = new StringBuilder();
-        buf.append(String.format("{{id:%d%s name:'%s'%s rowCount:%d%s version:%d%s owner:%s(%d)%s source:%s%s accessLevel:%s%s " +
-                        "readOnly:%s%s link:%s%s ganttEnabled:%s%s dependEnabled:%s%s resMgmntEnabled:%s%s favorite:%s%s}%n",
-                sheet.getId(), LINE_DELIMITER, sheet.getName(), LINE_DELIMITER, sheet.getTotalRowCount(), LINE_DELIMITER,
-                sheet.getVersion(), LINE_DELIMITER, sheet.getOwner(), sheet.getOwnerId(), LINE_DELIMITER,
-                sheet.getSource(), LINE_DELIMITER, sheet.getAccessLevel(), LINE_DELIMITER, sheet.getReadOnly(),
-                LINE_DELIMITER, sheet.getPermalink(), LINE_DELIMITER, sheet.getGanttEnabled(), LINE_DELIMITER,
-                sheet.getDependenciesEnabled(), LINE_DELIMITER, sheet.getResourceManagementEnabled(), LINE_DELIMITER,
-                sheet.getFavorite(), LINE_DELIMITER));
-        if (options != null && options.length > 0) {
-            for (ToStringOptions option : options) {
-                if (option != null) {
-                    option.appendData(buf, sheet, client);
-                }
-            }
-        }
-
-        buf.append("}");
-        return buf;
-    }
-
-    public static Row addRow(@NonNull SheetsApi client, long sheetId, @NonNull Supplier<Row> rowProvider) throws ApiException {
-        List<Row> rows = addRows(client, sheetId, Collections.singletonList(rowProvider.get()), null);
+    public static Row addRow(@NonNull ApiClient client, long sheetId, @NonNull Supplier<Row> rowProvider) throws ApiException {
+        List<Row> rows = addRows(client, sheetId, List.of(rowProvider.get()), null);
         return rows.isEmpty() ? null : rows.get(0);
     }
-*/
-/*
 
     @NonNull
-    public static List<Row> addRows(@NonNull SheetsApi client, long sheetId, @NonNull List<Row> rowData,
-                                    Consumer<List<Row>> cb) throws ApiException {
-        List<Row> newRows = client.sheetResources().rowResources().addRows(sheetId, rowData);
-        Preconditions.checkState(rowData.size() == newRows.size(), "%s rows sent, only %s returned", rowData.size(),
-                newRows.size());
+    public static List<Row> addRows(@NonNull ApiClient client, long sheetId, @NonNull List<Row> rowData, Consumer<List<Row>> cb)
+            throws ApiException {
+        var response = new RowsApi(client).rowsAddToSheet(sheetId, null, null, null, null, rowData);
+        State.notNull(response, "null response from rowsAddToSheet");
+        var responseResult = State.notNull(response.getResult(), "null response result");
+        assertEqualRowCounts(rowData.size(), responseResult.size());
+        // convert the response result into Rows
+        List<Row> newRows = responseResult.stream().map(Converters::convert).collect(Collectors.toList());
         if (cb != null) {
             cb.accept(newRows);
         }
@@ -78,7 +47,7 @@ public class Sheets {
     }
 
     @NonNull
-    public static List<Row> updateRows(@NonNull SheetsApi client, long sheetId, @NonNull List<Row> rowData,
+    public static List<Row> updateRows(@NonNull ApiClient client, long sheetId, @NonNull List<Row> rowData,
                                        Consumer<List<Row>> cb) throws ApiException {
         // be sure all forbidden fields are cleared
         // InvalidRequestException: The attribute(s) row.rowNumber, row.createdAt, row.modifiedAt, row.sheetId are not allowed for this operation.
@@ -88,15 +57,17 @@ public class Sheets {
             row.setModifiedAt(null);
             row.setSheetId(null);
         });
-        List<Row> newRows = client.sheetResources().rowResources().updateRows(sheetId, rowData);
-        Preconditions.checkState(rowData.size() == newRows.size(), "%s rows sent, only %s returned", rowData.size(),
-                newRows.size());
+        var response = new RowsApi(client).updateRows(sheetId, null, null, null, null, rowData);
+        State.notNull(response, "null response from updateRows");
+        var responseResult = State.notNull(response.getResult(), "null response result");
+        assertEqualRowCounts(rowData.size(), responseResult.size());
+        List<Row> newRows = response.getResult().stream().map(Converters::convert).collect(Collectors.toList());
+        newRows.forEach(row -> row.setSheetId(sheetId));    // not in the update rows response?
         if (cb != null) {
             cb.accept(newRows);
         }
         return newRows;
     }
-*/
 
     @NonNull
     public static String columnInfo(@NonNull Sheet sheet) {
@@ -123,25 +94,26 @@ public class Sheets {
     static final List<Integer> ALL_ROW_NUMBERS = null;
     static final OffsetDateTime ROWS_MODIFIED_SINCE = OffsetDateTime.now();
 
-    public static Sheet getWholeSheet(@NonNull SheetsApi client, long sheetId) throws ApiException {
-        return client.getSheet(sheetId, null, null, ALL_SHEET_INCLUSIONS, NO_SHEET_EXCLUSIONS, ALL_COLUMN_IDS,
-                        ALL_FILTERS, null, null, NO_PAGE_SIZE_LIMIT, ALL_PAGE_NUMBERS, null, ALL_ROW_IDS, ALL_ROW_NUMBERS, 
-                        ROWS_MODIFIED_SINCE)
+    public static Sheet getWholeSheet(@NonNull ApiClient client, long sheetId) throws ApiException {
+        return new SheetsApi(client)
+                .getSheet(sheetId, null, null, ALL_SHEET_INCLUSIONS, NO_SHEET_EXCLUSIONS, ALL_COLUMN_IDS, ALL_FILTERS, null, null,
+                        NO_PAGE_SIZE_LIMIT, ALL_PAGE_NUMBERS, null, ALL_ROW_IDS, ALL_ROW_NUMBERS, ROWS_MODIFIED_SINCE)
                 .getSheet();
     }
 
-    public static Sheet getSheetNoRows(@NonNull SheetsApi client, long sheetId) throws ApiException {
+    public static Sheet getSheetNoRows(@NonNull ApiClient client, long sheetId) throws ApiException {
         // FIXME - need to support a comma-sep list of enum values; not just a single value (or null)
-        return client.getSheet(sheetId, null, null, List.of(SheetInclude.COLUMN_TYPE), List.of(SheetExclude.LINK_IN_FROM_CELL_DETAILS),
+        return new SheetsApi(client)
+                .getSheet(sheetId, null, null, List.of(SheetInclude.COLUMN_TYPE), List.of(SheetExclude.LINK_IN_FROM_CELL_DETAILS),
                         ALL_COLUMN_IDS, ALL_FILTERS, null, null, NO_PAGE_SIZE_LIMIT, ALL_PAGE_NUMBERS, null, ALL_ROW_IDS, ALL_ROW_NUMBERS, null)
                 .getSheet();
     }
 
     @NonNull
-    public static List<SheetListingDataInner> findByName(@NonNull SheetsApi client, @NonNull String sheetName) throws ApiException {
-        return Objects.requireNonNull(
-                        client.listSheets(null, null, true, null, false, null, null).getData())
-                .stream()
+    public static List<SheetListingDataInner> findByName(@NonNull ApiClient client, @NonNull String sheetName)
+            throws ApiException {
+        return new SheetsApi(client)
+                .listSheets(null, null, true, null, false, null, null).getData().stream()
                 .filter(sheet -> sheetName.equals(sheet.getName()))
                 .collect(Collectors.toList());
     }
@@ -149,17 +121,18 @@ public class Sheets {
     /**
      * remove all the rows from the specified sheet
      */
-/*
-    public static void clearRows(@NonNull SheetsApi client, @NonNull Sheet sheet) throws ApiException {
+    public static void clearRows(@NonNull ApiClient client, @NonNull Sheet sheet) throws ApiException {
         // check if sheet already has no rows;
         // otherwise this call fails with "InvalidRequestException: A required parameter is missing from your request: ids."
-        Set<Long> rowIds = sheet.getRows().stream().map(Row::getId).collect(Collectors.toSet());
+        List<Long> rowIds = sheet.getRows().stream().map(Row::getId).collect(Collectors.toList());
         if (!rowIds.isEmpty()) {
-            client.sheetResources().rowResources().deleteRows(sheet.getId(), rowIds, true);
+            new RowsApi(client).deleteRows(sheet.getId(), rowIds, true);
         }
     }
-*/
 
+    private static void assertEqualRowCounts(int expected, int actual) {
+        State.isTrue(expected == actual, "%s rows sent, only %s returned", expected, actual);
+    }
 /*
     @NonNull
     public static Sheet copyAndRefresh(@NonNull SheetsApi client, @NonNull Sheet original,
@@ -178,46 +151,6 @@ public class Sheets {
     @Deprecated // use Rows.clearLocations(Row) instead
     public static void clearLocations(@NonNull Row row) {
         Rows.clearLocations(row);
-    }
-*/
-
-/*
-    public enum ToStringOptions {
-        WITH_COLUMNS {
-            @Override
-            void appendData(@NonNull StringBuilder buf, @NonNull Sheet sheet, @NonNull SheetsApi client) throws ApiException {
-                ColumnsApi columnsApi = new ColumnsApi(ApiClients.getDefaultClient());
-                var columnPage = columnsApi.columnsListOnSheet(sheet.getId(), null, Constants.ALL_PAGES);
-                List<Column> columns = columnPage.getData();
-                buf.append(String.format("%ncolumns:{num:%d data:{%n", columns.size()));
-                for (Column col : columns) {
-                    buf.append(col.getId()).append(":'").append(col.getTitle()).append("',").append(LINE_DELIMITER);
-                }
-                buf.append("}}");
-            }
-        },
-        WITH_ROW_CONTENT {
-            @Override
-            void appendData(@NonNull StringBuilder buf, @NonNull Sheet sheet, @NonNull SheetsApi client) throws ApiException {
-                buf.append("{rows:{");
-                for (Row row : sheet.getRows()) {
-                    buf.append(row.getId()).append(":[");
-                    for (Cell cell : row.getCells()) {
-                        buf.append(cell.getValue()).append(',').append(LINE_DELIMITER);
-                    }
-                    buf.append("],\n");
-                }
-                buf.append("}}\n");
-            }
-        }, ALL {
-            @Override
-            void appendData(@NonNull StringBuilder buf, @NonNull Sheet sheet, @NonNull SheetsApi client) throws ApiException {
-                WITH_COLUMNS.appendData(buf, sheet, client);
-                WITH_ROW_CONTENT.appendData(buf, sheet, client);
-            }
-        };
-
-        abstract void appendData(@NonNull StringBuilder buf, @NonNull Sheet sheet, @NonNull SheetsApi client) throws ApiException;
     }
 */
 }